Move config to a designated config folder.
